New Services Branch
The services folder will serve to perform four major operations. 
First, it checks that the information given to the database is correct: Checks if the fields are empty or not. 
Second, it passes the new account information to the database where it will be given an AccountID. 
Third, it will request this new AccountID back from the database and then create a new JSON object that contains this AccountID and the original information. 
Fourth, it will pass this new JSON object to the View Layer, where it is stored with the associated button.