Create exceptions.js
These are exceptions used in the validation step of add account/record. We will also need to update these when we have a better idea how a storage error can arise. A storage error must be thrown in the storage layer, so Pratik may know more about which errors are to be considered.